change type of Ppvz_spp_prc of MonthDetailSale struct

+++ types/month_detail_sale.go
@@ -32,7 +32,7 @@ type MonthDetailSale struct {
 	Product_discount_for_report float64         `json:"product_discount_for_report"` // Согласованный продуктовый дисконт
 	Supplier_promo              float64         `json:"supplier_promo"`              // Промокод
 	Rid                         uint64          `json:"rid"`                         // Rid
-	Ppvz_spp_prc                uint64          `json:"ppvz_spp_prc"`
+	Ppvz_spp_prc                float64         `json:"ppvz_spp_prc"`
 	Ppvz_kvw_prc_base           float64         `json:"ppvz_kvw_prc_base"`
 	Ppvz_kvw_prc                float64         `json:"ppvz_kvw_prc"`
 	Ppvz_sales_commission       float64         `json:"ppvz_sales_commission"`
